Setup the project to "build" joosc
We do not need any setup at the moment, so the build phase is just
smoke and mirrors.

If we do need to add anything later on, we wil just script it in
the rake task 'marmoset' that I created. Personall, I can't stand
Makefile syntax, so the Makefile just runs the 'marmoset' rake
task to do the actual work.