Enable the #link_identifiers step for name resolution
At this point, ever single entity, AST node, and token, will get a chance
to respond to the #link_identifiers method, in a recursive manner, from
the root compilation unit down to the terminal tokens in an AST.

The strategy for actually doing resolution will be to override the
default #link_identifiers method for the relevant classes so that
linking can actually occur.

All AST nodes respond to #scope, which will return the correct Joos::Scope
for the AST node. You can then use the scope to resolve names.