Add an assertion that it is safe to wrap QualifiedIdentifiers in casts
According to GH-88, it might not be a cast! However, it seems that our
grammar is such that the non-casting case will become a Primary instead
of a plain Term, and so this weeder check and transform is still valid.

In the case that it is not, we will find out through testing.