Merge rake one task functionality into Joos::Compiler interactive methods
So now you can do this:

Joos::Compiler.load_dir_with_stdilb 'Je_duplicateMethod'

And it will figure out which assignment the test belongs to and if it
needs to load an entire directory or a single file. If the test name is
ambiguous, you can prefix the test name with "a5/" or whatever test
suite the test actually belongs to.