Enable type checking
As discussed earlier, we are rolling the type checking and name
resolution phases into one pass. One reason to do this is because we
are forced to resolve types before we can resolve method selectors, but
we need to resolve method types in order to resolve some other types.

So, the process is simple: recursively resolve and type check AST nodes,
and if the node happens to be a QualifiedIdentifier or Selector node,
then we can resolve the name after the recursive call returns.