Make Void types respond to #eql? and #hash correctly
So that they can take part in the hierarchy checking that has been setup.

The real implication of this is that we want to use an explicit object
to represent void instead of nil. The nil value is what we have if we
have not resolved the type of something, and does not itself respond
to the Type API.