Properly resolve GH-107
Since it is grammatically impossible to allocate a basic type unless it
is an array, we only need to check that the type is not abstract, or
for array that the type the array is wrapping is not abstract.

Why check arrays? Well, you can't call the default constructor on a class
which has no constructors and cannot be allocated.