Add Class#type and Interface#type
Since Java does not really do metaprogramming in the same way as a
language like Ruby (i.e. same namespace for static and non-static
methods), the type of a class/interface is itself, and the caller
will just have to know for themselves if they want a static or
non-static method/field.