Array assignability is more strict for primitives
And there are special considerations for char being used in place of
a short or byte. This seems to just be Java weirdness. The byte issue
makes sense for full Java because Java uses UTF-16, but I'm not sure
why we are not allowed to be assignment compatible with shorts...