Rescope blocks after every single local variable declaration
This eliminates all but one case forward reference for local variables.
The question that you then ask is if a variable initializer can refer
to a field with the same name, what happens?

According to javac, we should link to the variable we are declaring,
and then throw a "variable might not be initialized" error. Translating
to Joos, this means that the declaration shadows the field when we go
to check the initializer, and we then have a forward reference. Boom!